requirements-dev.txt
black
click
coverage
pre-commit
pygame
pylint
pytest
pytest-benchmark
pytest-black
pytest-cov
pytest-pylint
python-chess
reconchess
tqdm
twine
